ALX S.E c programming practical

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ double a;
 float b;
 char c;
 
-printf("the total size of variable q is equal to %d\n", sizeof(q));
+printf("the total size of variable q is equal to %lu\n", (unsigned long)sizeof(q));
 printf("the tatal size of double a is %lu\n", (unsigned long)sizeof(a));
 printf("the total size of float b is %lu\n", (unsigned long)sizeof(b));
 printf("the total size of char c is %lu\n", (unsigned long)sizeof(c));
